feat(pet): add pet component

+++ wowPets/wow-pets-core/src/petDimension/petDimensionService.js
@@ -1,4 +1,5 @@
 const petDimensionDao = require('./petDimensionDao')
+const {utils} = require('../other/utils')
 
 const petDimensionService = {
   insert(petDimension) {
@@ -6,7 +7,17 @@ const petDimensionService = {
     return result
   },
   delete(_id) {
-    return petDimensionDao.deleteOne(_id)
+    let result,_ids=[]
+    _id.includes(',') ? _ids = _id.split(',') : _ids.push(_id)
+    if(!utils.checkId(_ids)){
+      return utils.createIdErrorMsg()
+    }
+    if(_ids.length > 1){
+      result = petDimensionDao.deleteMany(_ids)
+    }else {
+      result = petDimensionDao.deleteOne(_id)
+    }
+    return result
   },
   update(petDimension) {
     return petDimensionDao.updateOne(petDimension)
