+++ wowPets/wow-pets-core/src/quality/qualityService.js
@@ -1,4 +1,5 @@
 const qualityDao = require('./qualityDao')
+const {utils} = require('../other/utils')
 
 const qualityService = {
   insert(quality) {
@@ -6,7 +7,17 @@ const qualityService = {
     return result
   },
   delete(_id) {
-    return qualityDao.deleteOne(_id)
+    let result,_ids=[]
+    _id.includes(',') ? _ids = _id.split(',') : _ids.push(_id)
+    if(!utils.checkId(_ids)){
+      return utils.createIdErrorMsg()
+    }
+    if(_ids.length > 1){
+      result = qualityDao.deleteMany(_ids)
+    }else {
+      result = qualityDao.deleteOne(_id)
+    }
+    return result
   },
   update(quality) {
     return qualityDao.updateOne(quality)
